requirements.txt
python-jose[cryptography]==3.3.0

fastapi==0.110.0
uvicorn[standard]==0.27.1

# MongoDB
beanie==1.25.0
motor==3.3.2
pymongo==4.6.1

# PDF Processing
pypdf==4.2.0

# Embeddings
sentence-transformers==2.6.1
torch==2.2.1
numpy==1.26.4

# LLM (Generator)
transformers==4.38.2
accelerate==0.27.2

# Validation
pydantic==2.6.3
pydantic-settings==2.2.1

# Optional but Recommended
python-multipart==0.0.9
